Code doesn't work very well anymore.
The zoomToLocation doesn't seem to be interfering with resizeMap, it
seems that resizeMap can be called too many times to quickly. Try
spamming the back button after loading the map many times and you will
see what I mean.

I was considering placing some sort of timer on it so that it doesn't
run too often.