Finished templitizing application.
All of the text (except for links) has been converted into variables
that are set on the config.js page. I also simplified the code that
sorts events by day (for the event schedule page).

The documentation has been updated to reflect these changes.

Known bugs: Some of the dividers on the event schedule page aren't
displaying properly (for both Friday and Saturday).